+++ changelog.md
@@ -5,6 +5,7 @@
 - Access component's html content with `vm.$html` (rather than `vm.$el.outerHTML`)
 - `vm.$find` allows you to search a component's html for matching elements or component instances.
 - `vm.$contains` tests whether a component contains another component or a html element.
+- slots option lets you easily insert slot content into a component. This is the equivalent of manually inserting slots with the `innerHTML` property.
 
 ## 0.2.0
 - stubComponents flag automatically stubs all *known* components